move common undelegate stuff to internal/
