+++ src/processor/fast/internal/mod.rs
@@ -1,3 +1,5 @@
 mod commit_finalize_internal;
+mod undelegate_internal;
 
 pub(crate) use commit_finalize_internal::*;
+pub(crate) use undelegate_internal::*;
